Moved method and fixed error
